AArch64 support
This patch will require a little rework before it gets submitted.

1) I don't know if the Arduino team will simply edit the existing binary
dependencies in place and hence NOT change the filename in the way I have
below (libastylej-2.05.1-3.zip -> libastylej-2.05.1-4.zip), or will indeed
upload to new a filename.

2) Since the Arduino team will build and distribute all the binaries, the
SHA values in the patch below will have to be recalculated once the actual
files are available. The values below are for my locally built versions.